requirements.txt
dvc
dvc[gdrive]
sklearn
pandas
pyyaml
pytest
tox
flake8
flask
gunicorn